RavenDB-20652 Moving abstracts and types of SortingMatches into `Meta` directory.
