+++ src/Corax/Corax.csproj
@@ -38,6 +38,7 @@
   </ItemGroup>
 
 
+
   <PropertyGroup Condition=" '$(Configuration)' == 'Validate' ">
     <DefineConstants>$(DefineConstants);VALIDATE;VALIDATE_PAGES</DefineConstants>
     <Optimize>true</Optimize>
