+++ src/Corax/Queries/SortingMatches/Meta/MatchCompareFieldType.cs
@@ -1,0 +1,12 @@
+namespace Corax.Queries.SortingMatches.Meta;
+
+public enum MatchCompareFieldType : ushort
+{
+    Sequence,
+    Integer,
+    Floating,
+    Score,
+    Alphanumeric,
+    Spatial,
+    Random
+}
