+++ src/Corax/Queries/SortingMatches/SortingMultiMatch.Helpers.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Runtime.InteropServices;
+using Corax.Queries.SortingMatches.Meta;
 using Corax.Utils;
 using Sparrow;
