+++ src/Corax/Queries/SortingMatches/SortingMultiMatch.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Runtime.CompilerServices;
+using Corax.Queries.SortingMatches.Meta;
 using Corax.Utils;
 using Sparrow;
 using Sparrow.Server;
