+++ test/FastTests/Corax/Bugs/FacetIndexingRepro.cs
@@ -2,6 +2,7 @@ using System.IO;
 using Corax;
 using Corax.Mappings;
 using Corax.Queries.SortingMatches;
+using Corax.Queries.SortingMatches.Meta;
 using Corax.Utils;
 using FastTests.Voron;
 using Sparrow.Server;
