+++ test/FastTests/Corax/OrderBySorting.cs
@@ -4,6 +4,7 @@ using System.Text;
 using Corax;
 using Corax.Mappings;
 using Corax.Queries.SortingMatches;
+using Corax.Queries.SortingMatches.Meta;
 using Corax.Utils;
 using FastTests.Voron;
 using Sparrow.Server;
